adapter refactor + kafka simple test
