+++ src/main/java/it/gov/pagopa/rtd/ms/enrolledpaymentinstrument/adapters/event/KafkaAdapter.java
@@ -1,8 +1,9 @@
-package it.gov.pagopa.rtd.ms.enrolledpaymentinstrument.adapters;
+package it.gov.pagopa.rtd.ms.enrolledpaymentinstrument.adapters.event;
 
 import it.gov.pagopa.rtd.ms.enrolledpaymentinstrument.application.SomethingService;
 import java.util.function.Consumer;
 import lombok.extern.slf4j.Slf4j;
+import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.messaging.Message;
