+++ src/test/resources/application-test.yml
@@ -2,7 +2,37 @@ logging:
   level:
     root: INFO
 
+test.kafka.topic: "embedded-test-topic"
+
 spring:
   config:
     activate:
       on-profile: test
+
+  kafka:
+    template.default-topic: ${test.kafka.topic}
+    bootstrap-servers:
+      - ${spring.embedded.kafka.brokers}
+    producer:
+      key-serializer: org.apache.kafka.common.serialization.StringSerializer
+      value-serializer: org.apache.kafka.common.serialization.StringSerializer
+
+  cloud:
+    stream:
+      bindings:
+        enrolledPaymentInstrumentConsumer-in-0:
+            destination: ${test.kafka.topic}
+            group: rtd-enrolled-payment-instrument-consumer-group
+            binder: kafka-binder
+
+      binders:
+        kafka-binder:
+          type: kafka
+          environment:
+            spring.cloud.stream.kafka:
+              binder:
+                auto-create-topics: false
+                brokers: ${spring.embedded.kafka.brokers}
+                consumerProperties:
+                  key.deserializer: org.apache.kafka.common.serialization.StringDeserializer
+                  value.deserializer: org.apache.kafka.common.serialization.StringDeserializer
